requirements.txt
tqdm
llama-cpp-python
requests
pyperclip
questionary
